+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-03-18
+## [Unreleased] - 2022-03-21
 
 Contains breaking changes, including increasing minimum supported Java version to Java 17.
 
@@ -12,6 +12,8 @@ Contains breaking changes, including increasing minimum supported Java version t
 
 ### Changed
 * Minimum supported Java version is now Java 17.
+* Refactored all classes in package org.cicirello.math.rand to use Java 17's new
+  RandomGenerator interface.
 * Migrated test cases to JUnit Jupiter 5.8.2.
 
 ### Deprecated